requirements.txt
# Core dependencies
boto3>=1.34.0
botocore>=1.34.0
PyJWT>=2.8.0

# AWS CDK
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.90.0
moto>=4.2.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
